Add files via upload
Capstone Project (Aug/Sep -21)
Description : 
Predicting Corporate Company Bankruptcy based on historical data of Polish Companies.
Bankruptcy is a serious issue in every economy and entails a series of connected Financial institutions ,individuals and investors.
The objective of the model was to better equip investors both global and local, give financial institutions a broader picture of the company before investing/lending money.
There are widespread repercussion when a company goes bankrupt and financial turmoil can last for years. To avoid the distress situation one particular dimension is an informed decision about the company. And so is been ensured while model building.

Data description:
The dataset is about bankruptcy prediction of Polish companies. The data was collected from Emerging Markets Information Service (EMIS), which is a database containing information on emerging markets around the world. The bankrupt companies were analyzed in the period 2000-2012, while the still operating companies were evaluated from 2007 to 2013.
The dataset had 63 features input features and 1 target feature with 43405 rows leading to a binary class classification problem. (1:Bankrupt , 0:Not Bankrupt) 
These features were financial ratios so that the true Financial statements of the companies can only be assumed and privacy of data is maintained.
Also the size and domain of companies was not fixed leading to some unsupervised learning.
Extensive EDA was done. Imputing Missing values, finding out correlation of features with target variable and statistical analysis.

Model Building:
The class labels were imbalanced so SMOTE and SMOTETomek was done.
Models used : LogisticRegression, DecisionTreeClassifier, RandomForestClassifier, GradientBoostingClassifier, XGBoostClassifier

Evaluation Metric score used : f1_micro
Also Model building was done to reduce type II error (False Negatives)
Best Model : XGBoostClassifier  f1_micro = 98.58%

Inferences and Appendix:
Computation power restricted some of the model building. FNs can further be reduced.
In India many bankruptcy cases are pending under the The insolvency and Bankruptcy Code,2016(IBC) and as of now(Sep-2021)  cases are taking around 593 days to resolve.
It is vital to act before hand and take measures to avoid the financial turmoil. 
Also a True Audit report of company is detrimental for this model to work. The Financial information of the company given must be true and trustworthy.
Otherwise the model would fail because of manipulated financial statements as it happened in the case of Evergrande ,China. Where true information was disclosed late.
Another Factor which might affect this model would be government policies which may triggered a chain of unwanted events.